+++ devtools/projects/ng-devtools-backend/src/lib/directive-forest/identity-tracker.spec.ts
@@ -14,11 +14,15 @@ describe('IdentityTracker', () => {
   beforeEach(() => {
     (IdentityTracker as any).instance = undefined;
     tracker = IdentityTracker.getInstance();
+    (window as any).ng = {
+      getComponent: () => {},
+    };
   });
 
   afterEach(() => {
     (IdentityTracker as any).instance = undefined;
     document.querySelectorAll('[ng-version]').forEach((el) => el.remove());
+    delete (window as any).ng;
   });
 
   function seedDirective(opts: {dir: object; id: number; isComponent?: boolean}): void {
