+++ packages/core/primitives/devtools/index.ts
@@ -9,3 +9,8 @@
 export {ProfilerEvent, type Profiler} from './src/profiler_types';
 export {Framework} from './src/framework';
 export type {DevtoolsToolDiscoveryEvent, ToolDefinition, ToolGroup} from './src/tool_definitions';
+export type {
+  DebugSignalGraph,
+  DebugSignalGraphEdge,
+  DebugSignalGraphNode,
+} from './src/debug_signal_graph';
